[BUG] Fix dynamic_stacking failure
The following error is raised when running autogluon-assistant with dynamic_stacking enabled.
The reason for the error is that num_stack_levels is set in the following line at predictor.py:

`predictor_fit_kwargs["num_stack_levels"] = 1`

but the predictor_fit_kwargs is managed by OmegaConf, hence it is required to add the key in config.
Setting the default value to 0 here. Also adds ability to change num_stack_levels value from config.

omegaconf.errors.ConfigKeyError: Key 'num_stack_levels' is not in struct
    full_key: autogluon.predictor_fit_kwargs.num_stack_levels
    object_type=dict